requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.4.0
brotli==1.2.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
choicelib==0.1.5
colorama==0.4.6
cryptography==46.0.3
frozenlist==1.8.0
greenlet==3.3.0
idna==3.11
jeepney==0.9.0
msgspec==0.19.0
multidict==6.7.0
mutagen==1.47.0
playwright==1.57.0
propcache==0.4.1
pycparser==2.23
pycryptodomex==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.0
python-dotenv==1.2.1
requests==2.32.5
SecretStorage==3.5.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
vbml==1.1.post1
vkbottle==4.6.2
vkbottle-types==5.199.99.12
websockets==15.0.1
yarl==1.22.0
yt-dlp==2025.12.8
yt-dlp-ejs==0.3.2
